requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
Flask-Mail==0.10.0
Werkzeug==3.1.3
python-dotenv==1.1.0
PyJWT==2.10.1
numpy==2.2.5
scikit-learn==1.6.1
scipy==1.15.2
joblib==1.4.2
pandas>=2.0.0
pymongo>=4.0.0
requests>=2.26.0
gunicorn>=20.1.0
pymysql>=1.0.2
python-dateutil>=2.8.2
tqdm>=4.65.0
jieba>=0.42.1
snownlp>=0.12.3 